requirements.txt
Django==4.0
djangorestframework==3.11.0
SQLAlchemy==1.4.40
sqlparse==0.3.1
whitenoise==5.2.0
